refactor(agents): split pilot.ts into modular components (Issue #697)
Split the 895-line pilot.ts into smaller, more maintainable modules
under src/agents/pilot/ directory.

## Changes

| File | Lines | Description |
|------|-------|-------------|
| `pilot/types.ts` | ~84 | Type definitions (PilotCallbacks, PilotConfig, MessageData) |
| `pilot/message-builder.ts` | ~217 | Message building utilities (buildEnhancedContent, etc.) |
| `pilot/pilot.ts` | ~642 | Core Pilot class (reduced from 895 lines) |
| `pilot/index.ts` | ~14 | Module exports |
| `pilot.ts` (root) | ~9 | Re-export for backward compatibility |

Main file reduced by ~250 lines (~28% reduction)

## Key Changes

- **types.ts**: Extracted PilotCallbacks, PilotConfig, MessageData interfaces
- **message-builder.ts**: Extracted buildEnhancedContent, buildToolsSection, buildAttachmentsInfo
- **Backward Compatibility**: Root pilot.ts re-exports from pilot/ module

## Test Results

- All 159 agent tests pass
- TypeScript compilation successful

Fixes #697
